Agregados permisos en fx y sort a tables

+++ fichaube/permisos/views.py
@@ -29,36 +29,46 @@ from permisos.models import Permiso
 
 @login_required()
 def cederPermiso(request, id_alumno=None):
-    template = "ceder_permiso.html"
 
-    if request.method == 'GET':
-        ficha = Ficha.objects.get(alumno_id = id_alumno)
-        querys = (Q( profesional= True) | Q(asistente_social = True))
-        profesionales = Usuario.objects.filter(querys)
-        profesionalesSinPermiso = profesionales.exclude(permiso__ficha__id = ficha.id)
-        return render(request, template, {'usuarios': profesionalesSinPermiso, 'ficha':ficha})
+    #PERMISOS
+    if request.user.groups.filter(name__in=['Mantenedor']).exists():
+        return HttpResponseRedirect(reverse("home"))
+    elif request.user.groups.filter(name__in=['Profesional']).exists():
+        return HttpResponseRedirect(reverse("home"))
+    elif request.user.groups.filter(name__in=['Asistente Social']).exists():
+        return HttpResponseRedirect(reverse("home"))
+    else:
 
-    if request.method == 'POST':
-        try:
-            rutUsuario = request.POST.get('inputRutProfesional')
-            print(rutUsuario)
-            print(id_alumno)
-            if rutUsuario:
-                permiso = Permiso()
-                permiso.ficha = Ficha.objects.get(alumno_id = id_alumno)
-                permiso.usuario = Usuario.objects.get(rut = rutUsuario)
-                permiso.save()
-                del permiso
+        template = "ceder_permiso.html"
 
-                messages.success(request, '¡Permiso cedido con éxito!')
+        if request.method == 'GET':
+            ficha = Ficha.objects.get(alumno_id = id_alumno)
+            querys = (Q( profesional= True) | Q(asistente_social = True))
+            profesionales = Usuario.objects.filter(querys)
+            profesionalesSinPermiso = profesionales.exclude(permiso__ficha__id = ficha.id)
+            return render(request, template, {'usuarios': profesionalesSinPermiso, 'ficha':ficha})
+
+        if request.method == 'POST':
+            try:
+                rutUsuario = request.POST.get('inputRutProfesional')
+                print(rutUsuario)
+                print(id_alumno)
+                if rutUsuario:
+                    permiso = Permiso()
+                    permiso.ficha = Ficha.objects.get(alumno_id = id_alumno)
+                    permiso.usuario = Usuario.objects.get(rut = rutUsuario)
+                    permiso.save()
+                    del permiso
+
+                    messages.success(request, '¡Permiso cedido con éxito!')
+                    return HttpResponseRedirect(reverse("alumnos:verAlumno", args=[id_alumno]))
+
+                messages.error(request,'¡No se elegió usuario!')
                 return HttpResponseRedirect(reverse("alumnos:verAlumno", args=[id_alumno]))
 
-            messages.error(request,'¡No se elegió usuario!')
-            return HttpResponseRedirect(reverse("alumnos:verAlumno", args=[id_alumno]))
-
-        except Exception as e:
-            messages.error(request,'¡No se pudo ceder permisos al usuario!')
-            return HttpResponseRedirect(reverse("alumnos:verAlumno", args=[id_alumno]))
+            except Exception as e:
+                messages.error(request,'¡No se pudo ceder permisos al usuario!')
+                return HttpResponseRedirect(reverse("alumnos:verAlumno", args=[id_alumno]))
 
 
 
@@ -68,15 +78,24 @@ def cederPermiso(request, id_alumno=None):
 @login_required()
 def revocarPermiso(request, id_permiso=None, id_alumno=None):
 
-    if request.method == 'GET':
-        try:
-            permiso = Permiso.objects.get(id = id_permiso)
-            permiso.delete()
-            del permiso
+    #PERMISOS
+    if request.user.groups.filter(name__in=['Mantenedor']).exists():
+        return HttpResponseRedirect(reverse("home"))
+    elif request.user.groups.filter(name__in=['Profesional']).exists():
+        return HttpResponseRedirect(reverse("home"))
+    elif request.user.groups.filter(name__in=['Asistente Social']).exists():
+        return HttpResponseRedirect(reverse("home"))
+    else:
 
-            messages.success(request, '¡Permiso revocado con éxito!')
-            return HttpResponseRedirect(reverse("alumnos:verAlumno", args=[id_alumno]))
+        if request.method == 'GET':
+            try:
+                permiso = Permiso.objects.get(id = id_permiso)
+                permiso.delete()
+                del permiso
 
-        except ObjectDoesNotExist:
-            messages.error(request,'ERROR - ¡No se pudo revocar el permiso correctamente!')
-            return HttpResponseRedirect(reverse("alumnos:verAlumno", args=[id_alumno]))
+                messages.success(request, '¡Permiso revocado con éxito!')
+                return HttpResponseRedirect(reverse("alumnos:verAlumno", args=[id_alumno]))
+
+            except ObjectDoesNotExist:
+                messages.error(request,'ERROR - ¡No se pudo revocar el permiso correctamente!')
+                return HttpResponseRedirect(reverse("alumnos:verAlumno", args=[id_alumno]))
